encapsulate snippet logic
