+++ src/core/buffer/index.ts
@@ -6,7 +6,7 @@ import { isThenable } from '../../lib/is-thenable'
  * The names of any Analytics instance methods that can be called pre-initialization.
  * These methods should exist statically on AnalyticsBrowser.
  */
-type PreInitMethodName =
+export type PreInitMethodName =
   | 'trackSubmit'
   | 'trackClick'
   | 'trackLink'
@@ -89,43 +89,8 @@ export interface PreInitMethodCall<
   reject: (reason: any) => void
 }
 
-const normalizeSnippetBuffer = (buffer: SnippetBuffer): PreInitMethodCall[] => {
-  return buffer.map(
-    ([methodName, ...args]) =>
-      ({
-        method: methodName,
-        resolve: () => {},
-        reject: console.error,
-        args,
-        called: false,
-      } as PreInitMethodCall)
-  )
-}
-
-type PreInitMethodParams<MethodName extends PreInitMethodName> = Parameters<
-  Analytics[MethodName]
->
-
-type SnippetWindowBufferedMethodCall<
-  MethodName extends PreInitMethodName = PreInitMethodName
-> = [MethodName, ...PreInitMethodParams<MethodName>]
-
-/**
- * A list of the method calls before initialization for snippet users
- * For example, [["track", "foo", {bar: 123}], ["page"], ["on", "ready", function(){..}]
- */
-type SnippetBuffer = SnippetWindowBufferedMethodCall[]
-
-/**
- * Fetch the buffered method calls from the window object and normalize them.
- */
-export const getSnippetWindowBuffer = (): PreInitMethodCall[] => {
-  const wa = window.analytics
-  const buffered =
-    // @ts-expect-error
-    (wa && wa[0] ? [...wa] : []) as SnippetBuffer
-  return normalizeSnippetBuffer(buffered)
-}
+export type PreInitMethodParams<MethodName extends PreInitMethodName> =
+  Parameters<Analytics[MethodName]>
 
 /**
  * Infer return type; if return type is promise, unwrap it.
